GDPR_anonymisation_package/requirements.txt
google-cloud-bigquery
Jinja2
google-auth
PyYAML